requirements.txt
requests==2.31.0
google-cloud-storage==2.14.0
google-cloud-bigquery==3.17.2
python-dotenv==1.0.1
tenacity==8.2.3

# Orchestration
prefect>=2.19,<3.0

# dbt
dbt-bigquery>=1.8,<2.0

# Dev / tests
pytest>=8.0
pytest-cov>=5.0
